# doc/phase1/swagger.yaml
swagger: "2.0"
info:
  description: "This is the api spec for Team SpaceWhales."
  version: "1.0.0"
  title: ""
  contact: 
    name: ""
  license:
    name: "<insert license here>"
host: "<unknown_domain_insert_here>"
basePath: "/space_whales"
...
schemes:
- "https"
paths:
...
  /users/manage:
    put:
      tags: 
      - "User"
      summary: "Create new user"
      description: ""
      operationId: "newUser"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NewUserRequest"
      responses:
        200:
          description: "Successful user creation"
        401:
          description: "Invalid account creation arguments"

    post:
      tags:
      - "User"
      summary: "Update a user's info" 
      description: ""
      operationId: "updateUser"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/UpdateUserRequest"
      responses:
        200:
          description: "Successful user update"
        401:
          description: "Invalid user update arguments"
        405:
          description: "Invalid token"

    delete:
      tags:
      - "User"
      summary: "Delete a user"
      description: ""
      operationId: "deleteUser"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/LoginToken"
      responses:
        200:
          description: "Successful user deletion"
        401:
          description: "Could not delete user"
        405:
          description: "Invalid token"
  /users/token/login:
    post:
      tags:
      - "User"
      summary: "Create and return user login token based on provided username and password hash"  
      description: ""
      operationId: "getUserToken"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/LoginRequest"
      responses:
        200:
          description: "Successful login"
          schema:
            $ref: "#/definitions/LoginToken"
        405:
          description: "Invalid login credentials"
  /users/token/invalidateToken:
    post:
      tags:
      - "User"
      summary: "Invalidate an existing token - primary use case is logoff."
      description: ""
      operationId: "invalidateToken"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/LoginToken"
      responses:
        200:
          description: "Successful invalidation"
        405:
          description: "Provided token is already expired"
        418:
          description: "Fake token"
  /users/token/validateToken:
    post:
      tags:
      - "User"
      summary: "Validates whether a given token is valid - primarily for inter-service communication"
      description: ""
      operationId: "authenticateToken"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/LoginToken"
      responses:
        200:
          description: "Token is valid"
        405:
          description: "Expired token"
        418:
          description: "Fake token"
  /users/token/validateToken/admin:
    post:
      tags:
      - "User"
      summary: "Validates whether a given token is a valid admin token"
      description: ""
      operationId: "authenticateAdminToken"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/LoginToken"
      responses:
        200:
          description: "Token is valid admin token"
        405:
          description: "Invalid token"
...
definitions:
  Username:
    type: "string"
    minLength: 8
    maxLength: 32
...
  LoginRequest:
    type: "object"
    properties:
      username:
        $ref: "#/definitions/Username"
      password:
        type: "string"
        description: "Password hashed by client"
  LoginToken:
    type: "object"
    properties:
      username:
        $ref: "#/definitions/Username"
      sessionToken:
        type: "string"
        minLength: 128
        maxLength: 256
      expiryTime:
        type: "string"
        format: "date-time"
...
  NewUserRequest:
    type: "object"
    properties:
      username:
        $ref: "#/definitions/Username"
      password:
        type: "string"
      email:
        type: "string"
        format: "email"
  UpdateUserRequest:
    type: "object"
    properties:
      loginToken:
        $ref: "#/definitions/LoginToken"
#idk what we actually want to let them change, so idk what more to put here